refactor: Separate getUpperCasePosition function
